add database config
